pyproject.toml
[tool.poetry]
name = "note"
version = "0.1.0"
description = "A simple note-taking application."
authors = ["Baiel <jyrgalbekuulubaiel@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
django-environ = "^0.11.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tools.isort]
py_version = 3.11
profile = 'django'
line_length = 120
multi_line_output = 2
known_first_party = ['note', 'core']
